feat: add username and email check in frontend
