+++ server/Dockerfile
@@ -4,10 +4,12 @@ WORKDIR /app
 COPY ./package.json yarn.lock ./
 COPY server/package.json ./server/package.json
 COPY packages/file-graph/package.json ./packages/file-graph/package.json
+COPY packages/utils/package.json ./packages/utils/package.json
 RUN yarn install --frozen-lockfile --production
 
 COPY ./server ./server
 COPY packages/file-graph ./packages/file-graph
+COPY packages/utils ./packages/utils
 COPY quests ./quests
 
 WORKDIR /app/server
